requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes==1.4.14
edge-tts==7.2.7
filelock==3.20.1
Flask==3.1.2
frozenlist==1.8.0
fsspec==2025.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
ollama==0.6.1
openai-whisper==20250625
pillow==12.0.0
propcache==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyperclip==1.11.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pyttsx3==2.99
pytweening==1.2.0
pywhatkit==5.4
pywin32==311
regex==2025.11.3
requests==2.32.5
soupsieve==2.8.1
SpeechRecognition==3.14.4
sympy==1.14.0
tabulate==0.9.0
tiktoken==0.12.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
wikipedia==1.4.0
yarl==1.22.0
